third_party/blink/web_tests/wpt_internal/js/source-phase-imports/invalid-module-types-throw-expected.txt

This is a testharness.js-based test.
[FAIL] Not supported source phase import types should throw a syntax error/
  assert_equals: expected "Uncaught SyntaxError: ./resources/test_module.js cannot be imported. Only Wasm module imports are supported in the source phase." but got "Uncaught SyntaxError: Unexpected identifier 'exportedNamesSource'"
Harness: the test ran to completion.
